+++ macroetym/main.py
@@ -2,17 +2,6 @@
 The Macro-Etymological Analyzer.
 Author: Jonathan Reeve, jonathan@jonreeve.com
 License: GPLv3
-
-This program looks up all the words in your text, using the Etymological
-Wordnet.
-
-I made the file included here, etymwn-smaller.tsv, by running these unix
-commands on the Etymological Wordnet:
-
-First, get only those entries with the relation "rel:etymology": grep
-    "rel:etymology" etymwn.tsv > etymwn-small.tsv Now we can remove the relation
-    column, since it's all "rel:etymology": cat etymwn-small.tsv | cut -f1,3 >
-    etymwn-smaller.tsv
 """
 
 from collections import Counter
@@ -22,7 +11,6 @@ import spacy
 from spacy.cli.download import download as spacy_download
 from spacy.tokens import Doc, Token
 from importlib.resources import files
-from pycountry import languages
 import click
 import csv
 import logging
@@ -34,8 +22,8 @@ import os
 import gzip
 from pathlib import Path
 from rich.progress import Progress
-import tempfile
-from . import datamaker
+import time
+from . import data_download
 
 # --- SpaCy Model Management & Custom Components ---
 
@@ -48,13 +36,7 @@ def etymology_lookup_component(doc):
     """
     SpaCy component to look up etymologies for tokens and store them in a custom attribute.
     """
-    lang_3 = doc.user_data.get("lang_3")
-    if not lang_3:
-        return doc
-    for token in doc:
-        if not token.is_stop and not token.is_punct and token.is_alpha:
-            word_obj = Word(token.lemma_.lower(), lang=lang_3)
-            token._.parent_languages = word_obj.parent_languages
+    # This component will be refactored to use the new database.
     return doc
 
 SPACY_MODEL_MAP = {
@@ -89,6 +71,8 @@ def load_spacy_model(lang: str):
         spacy_download(model_name)
         nlp = spacy.load(model_name)
 
+    nlp.max_length = 2000000
+
     if "etymology_lookup" not in nlp.pipe_names:
         nlp.add_pipe("etymology_lookup", last=True)
         
@@ -97,372 +81,297 @@ def load_spacy_model(lang: str):
 # --- End SpaCy Model Management & Custom Components ---
 
 
-# Parse the CSV file.
-etymdict = {}
-etymwn = files(__name__).joinpath('etymwn-smaller.tsv')
-with open(etymwn, encoding='utf-8') as csvfile:
-    csvreader = csv.reader(csvfile, delimiter='\t')
-    for line in csvreader:
-        if line[0] in etymdict:
-            etymdict[line[0]].append(line[1])
-        else:
-            etymdict[line[0]] = [line[1]]
-
-class LangList():
-    """
-    A class for language lists, that helps to count languages.
-    """
-    def __init__(self, langs):
-        self.langs = langs
-
-    def __repr__(self):
-        return str(self.langs)
-
-    @property
-    def stats(self):
-        """ Generates statistics about languages present in the list. """
-        counter = Counter(self.langs)
-        stats = {}
-        for lang in counter.keys():
-            stats[lang] = (counter[lang] / len(self.langs))*100
-        return stats
-
-
-class Word():
-    """
-    A word object, for looking up etymologies of single words.
-    """
-    def __init__(self, word, lang='eng', ignoreAffixes=True, ignoreCurrent=True):
-        self.lang = lang
-        self.word = word
-        self.ignoreAffixes = ignoreAffixes
-        self.ignoreCurrent = ignoreCurrent
-
-    def __repr__(self):
-        return '%s (%s)' % (self.word, self.lang)
-
-    def __str__(self):
-        return self.word
-
-    @staticmethod
-    def old_versions(language):
-        """
-        Returns a list of older versions of a language, such that given "eng"
-        (Modern English) it will return "enm" (Middle English). This is used
-        for filtering out current languages in the ignoreCurrent option of
-        parents() below.
-        """
-        lang_map = {'eng': ['enm'],
-                    'fra': ['frm', 'xno'],
-                    'dut': ['dum'],
-                    'gle': ['mga']}
-        # TODO: add other languages here.
-        return lang_map.get(language, [])
-
-    @property
-    def parents(self):
-        """ A wrapper for getParents"""
-        return self.getParents()
-
-    def getParents(self, level=0):
-        """
-        The main etymological lookup method.
-
-        ignoreAffixes will remove suffixes like -ly, so that the parent list
-        for "universally" returns "universal (eng)" instead of "universal
-        (eng), -ly (eng)."
-
-        ignoreCurrent will ignore etymologies in the current language and
-        slightly older versions of that language, so that it skips "universal
-        (eng)," and goes straight to the good stuff, i.e. "universalis (lat)."
-        Given a word in English, it will skip all other English and Middle
-        English ancestors, but won't skip Old English.
-        """
-        language = self.lang
-
-        # Finds the first-generation ancestor(s) of a word.
-        try:
-            raw_parent_list = etymdict[language + ": " + self.word]
-        except:
-            raw_parent_list = []
-        parent_list = [self.split(parent) for parent in raw_parent_list]
-        if self.ignoreAffixes:
-            parent_list = [p for p in parent_list if p.word[0] != '-']
-            parent_list = [p for p in parent_list if p.word[-1] != '-']
-        if self.ignoreCurrent:
-            newParents = []
-            for parent in parent_list:
-                if (parent.lang == language or parent.lang in self.old_versions(language)) and level < 3:
-                    logging.debug('Searching deeper for word %s with lang %s' % (parent.word, parent.lang))
-                    for otherParent in parent.getParents(level=level+1):
-                        # Go deeper.
-                        newParents.append(otherParent)
-                else:
-                    newParents.append(parent)
-            parent_list = newParents
-        return parent_list
-
-    @property
-    def parent_languages(self):
-        """ Get the parent languages of a word."""
-        parent_langs = []
-        for parent in self.parents:
-            parent_langs.append(parent.lang)
-        return LangList(parent_langs)
-
-    @property
-    def grandparents(self):
-        """ Get the grandparent words/languages of a word."""
-        return [Word(parent.word, lang=parent.lang).parents
-        for parent in self.parents]
-
-    @property
-    def grandparent_languages(self):
-        """ Get the grandparent languages for a word. """
-        grandparent_langs = []
-        for grandparent_list in self.grandparents:
-            for grandparent in grandparent_list:
-                grandparent_langs.append(grandparent.lang)
-        return LangList(grandparent_langs)
-
-    @staticmethod
-    def split(expression):
-        """ Takes and expression in the form 'enm: not' and returns
-        a Word object where word.lang is 'enm' and word.word is 'not'.
-        """
-        parts = expression.split(':')
-        return Word(parts[1].strip(), parts[0])
-
-class Text():
-    """ A container for texts, powered by SpaCy. """
-    def __init__(self, text, lang='eng', ignoreAffixes=True, ignoreCurrent=True):
-        self.lang = lang
-        self.ignoreAffixes = ignoreAffixes
-        self.ignoreCurrent = ignoreCurrent
-        
-        logging.debug('Initializing text with lang %s', lang)
-        if ignoreAffixes:
-            logging.debug('Ignoring affixes.')
-        if ignoreCurrent:
-            logging.debug('Ignoring current language and its middle variants.')
-            
-        nlp = load_spacy_model(lang)
-        
-        # Manually create the Doc, set user_data, and process the pipeline
-        doc = nlp.make_doc(text)
-        doc.user_data["lang_3"] = self.lang
-        for name, proc in nlp.pipeline:
-            doc = proc(doc)
-        self.doc = doc
-
-    langDict = {'Germanic': ['eng', 'enm', 'ang', 'deu', 'dut', 'nld', 'dum',
-                             'non', 'gml', 'yid', 'swe', 'rme', 'sco', 'isl',
-                             'dan'],
-                'Latinate': ['fra', 'frm', 'fro', 'lat', 'spa', 'xno', 'por',
-                             'ita'],
-                'Indo-Iranian': ['hin', 'fas'],
-                'Celtic': ['gle', 'gla'],
-                'Hellenic': ['grc'],
-                'Semitic': ['ara', 'heb'],
-                'Turkic': ['tur'],
-                'Austronesian': ['tgl', 'mri', 'smo'],
-                'Balto-Slavic': ['rus'],
-                'Uralic': ['fin', 'hun'],
-                'Japonic': ['jpn']}
-
-    def annotate(self):
-        """ Returns an annotated text in HTML format. """
-        html = ""
-        return html
-
-    def showMacroEtym(self):
-        for token in self.doc:
-            if token._.parent_languages:
-                print(token.text, token._.parent_languages)
-
-    def getStats(self, pretty=False):
-        stats_list = [token._.parent_languages.stats for token in self.doc if token._.parent_languages]
-        
-        stats = {}
-        for item in stats_list:
-            if len(item) > 0:
-                for lang, perc in item.items():
-                    if lang not in stats:
-                        stats[lang] = perc
-                    else:
-                        stats[lang] += perc
-        
-        if not stats:
-            return {}
-
-        allPercs = sum(stats.values())
-        for lang, perc in stats.items():
-            stats[lang] = ( perc / allPercs ) * 100
-
-        if pretty:
-            prettyStats = {}
-            for lang, perc in stats.items():
-                try:
-                    prettyLang = languages.get(alpha_3=lang).name
-                except:
-                    prettyLang = "Other Language"
-                prettyStats[prettyLang] = round(perc, 2) # rename the key
-            return prettyStats
-        return stats
-
-    def getFamily(self, language):
-        for family, children in self.langDict.items():
-            if language in children:
-                return family
-        return 'Other'
-
-    def getFamilyStats(self):
-        stats = self.getStats()
-        families = {}
-        for lang, perc in stats.items():
-            fam = self.getFamily(lang)
-            #print( fam, lang, perc) #debugging
-            if fam in families:
-                families[fam].append((lang, perc))
-            else:
-                families[fam] = [(lang, perc)]
-        return families
-
-    def compileFamilyStats(self, pad=True):
-        families = self.getFamilyStats()
-        totals = {}
-        for family, langs in families.items():
-            totals[family] = 0
-            for lang in langs:
-                totals[family] += lang[1]
-        # optionally add language families not represented by the text
-        if pad:
-            for fam in self.langDict:
-                if fam not in totals:
-                    totals[fam] = 0.0
-        return totals
-
-    @property
-    def stats(self):
-        return self.getStats()
-
-    def familyStats(self, pad=True):
-        return self.compileFamilyStats(pad)
-
-    @property
-    def prettyStats(self):
-        return self.getStats(pretty=True)
-
-    def printPrettyStats(self, filename):
-        d = {filename: self.prettyStats}
-        df = pd.DataFrame(d)
-        print(df)
-
-    def printCSVStats(self, filename):
-        d = {filename: self.prettyStats}
-        df = pd.DataFrame(d)
-        print(df.to_csv())
 
 @click.group()
 def cli():
     pass
+
+class Analysis():
+    DESIRED_SUBFAMILIES = {"Germanic", "Italic", "Indo-Aryan", "Iranian", "Hellenic", "Celtic", "Balto-Slavic"}
+    MANUAL_SUBFAMILY_MAP = {
+        'eng': 'Germanic',
+        'deu': 'Germanic',
+        'nld': 'Germanic',
+        'fra': 'Italic',
+        'ita': 'Italic',
+        'spa': 'Italic',
+        'por': 'Italic',
+        'ron': 'Italic',
+        'rus': 'Balto-Slavic',
+        'pol': 'Balto-Slavic',
+        'ukr': 'Balto-Slavic',
+        'ces': 'Balto-Slavic',
+        'hin': 'Indo-Aryan',
+        'urd': 'Indo-Aryan',
+        'ben': 'Indo-Aryan',
+        'fas': 'Iranian',
+        'ell': 'Hellenic',
+        'gle': 'Celtic',
+        'gla': 'Celtic',
+        'cym': 'Celtic',
+    }
+
+    def __init__(self, text, lang, conn):
+        self.text = text
+        self.lang = lang
+        self.conn = conn
+        self.c = self.conn.cursor()
+        self.language_counts = Counter()
+        self.family_counts = Counter()
+        self.glottolog_cache = {}
+        self.iso_to_glottocodes_list = {}
+        self.pycountry_cache = {}
+        self._load_glottolog_cache()
+        self._load_pycountry_cache()
+        self._analyze()
+
+    def _load_glottolog_cache(self):
+        self.c.execute("SELECT id, name, family, iso639p3code, family_path FROM glottolog")
+        for row in self.c.fetchall():
+            glottocode_id = row[0]
+            name = row[1]
+            family_id = row[2]
+            iso639p3code = row[3]
+            family_path = row[4]
+
+            self.glottolog_cache[glottocode_id] = {"name": name, "family_id": family_id, "iso639p3code": iso639p3code, "family_path": family_path}
+
+            if iso639p3code:
+                if iso639p3code not in self.iso_to_glottocodes_list:
+                    self.iso_to_glottocodes_list[iso639p3code] = []
+                self.iso_to_glottocodes_list[iso639p3code].append(glottocode_id)
+
+    def _load_pycountry_cache(self):
+        self.c.execute("SELECT alpha_2, alpha_3, name FROM pycountry_languages")
+        for row in self.c.fetchall():
+            if row[0]:
+                self.pycountry_cache[row[0]] = {"alpha_3": row[1], "name": row[2]}
+            if row[1]:
+                self.pycountry_cache[row[1]] = {"alpha_2": row[0], "name": row[2]}
+
+    def _get_language_name(self, iso_code):
+        if iso_code in self.pycountry_cache:
+            return self.pycountry_cache[iso_code]["name"]
+        return iso_code # fallback
+
+    def _analyze(self):
+        nlp = load_spacy_model(self.lang)
+        doc = nlp(self.text)
+        
+        full_lang_name = self._get_language_name(self.lang)
+
+        # Separate words into lowercase and proper nouns
+        lowercase_words = set()
+        proper_nouns = set()
+        for token in doc:
+            if not token.is_stop and not token.is_punct and token.is_alpha:
+                if token.pos_ == 'PROPN':
+                    proper_nouns.add(token.text)
+                else:
+                    lowercase_words.add(token.lemma_.lower())
+
+        etymologies = {}
+        batch_size = 500
+        
+        start_time = time.time()
+        
+        # Batch query for lowercase words
+        lemmas_list = list(lowercase_words)
+        for i in range(0, len(lemmas_list), batch_size):
+            batch = lemmas_list[i:i + batch_size]
+            placeholders = ','.join('?' for _ in batch)
+            query = f"""
+                SELECT w.word, s.etymology_chain
+                FROM words w
+                JOIN senses s ON w.id = s.word_id
+                WHERE w.word IN ({placeholders})
+            """
+            self.c.execute(query, batch)
+            for word, etymology_chain in self.c.fetchall():
+                if word not in etymologies:
+                    etymologies[word] = []
+                if etymology_chain:
+                    etymologies[word].append(etymology_chain)
+
+        # Batch query for proper nouns
+        proper_nouns_list = list(proper_nouns)
+        for i in range(0, len(proper_nouns_list), batch_size):
+            batch = proper_nouns_list[i:i + batch_size]
+            placeholders = ','.join('?' for _ in batch)
+            query = f"""
+                SELECT w.word, s.etymology_chain
+                FROM words w
+                JOIN senses s ON w.id = s.word_id
+                WHERE w.word IN ({placeholders})
+            """
+            self.c.execute(query, batch)
+            for word, etymology_chain in self.c.fetchall():
+                if word not in etymologies:
+                    etymologies[word] = []
+                if etymologies[word]:
+                    etymologies[word].append(etymology_chain)
+        
+        end_time = time.time()
+        
+        logging.info(f"Queried {len(lowercase_words) + len(proper_nouns)} unique words in {end_time - start_time:.2f} seconds.")
+
+        for token in doc:
+            if not token.is_stop and not token.is_punct and token.is_alpha:
+                word_to_lookup = token.text if token.pos_ == 'PROPN' else token.lemma_.lower()
+                if word_to_lookup in etymologies:
+                    for etymology_chain in etymologies[word_to_lookup]:
+                        langs = etymology_chain.split(' > ')
+                        self.language_counts.update(langs)
+                elif token.pos_ != 'PROPN':
+                    logging.debug(f"Word '{word_to_lookup}' not found in database for language '{full_lang_name}'.")
+        
+        for lang, count in self.language_counts.items():
+            family = self._get_family(lang)
+            if family:
+                self.family_counts[family] += count
+
+    def _get_family(self, lang_code):
+        # Handle proto-languages
+        if lang_code.endswith('-pro'):
+            return "Proto-Language"
+
+        # 1. Try manual mapping first
+        if lang_code in self.MANUAL_SUBFAMILY_MAP:
+            return self.MANUAL_SUBFAMILY_MAP[lang_code]
+        
+        # 2. Resolve lang_code to ISO 639-3
+        resolved_iso639_3_code = None
+        if lang_code in self.pycountry_cache: # Could be alpha_2 or alpha_3
+            if "alpha_3" in self.pycountry_cache[lang_code]:
+                resolved_iso639_3_code = self.pycountry_cache[lang_code]["alpha_3"]
+            else: # It's an alpha_3 code already
+                resolved_iso639_3_code = lang_code
+        elif len(lang_code) == 3: # Assume it's an ISO 639-3 code directly
+            resolved_iso639_3_code = lang_code
+        
+        if resolved_iso639_3_code in self.MANUAL_SUBFAMILY_MAP:
+            return self.MANUAL_SUBFAMILY_MAP[resolved_iso639_3_code]
+
+        # 3. Try to get Glottocode and then family path
+        canonical_glottocode = None
+        if resolved_iso639_3_code and resolved_iso639_3_code in self.iso_to_glottocodes_list:
+            canonical_glottocode = self.iso_to_glottocodes_list[resolved_iso639_3_code][0]
+
+        if not canonical_glottocode and lang_code in self.glottolog_cache:
+            canonical_glottocode = lang_code
+
+        if canonical_glottocode and canonical_glottocode in self.glottolog_cache:
+            family_path_str = self.glottolog_cache[canonical_glottocode].get("family_path")
+            if family_path_str:
+                family_path = family_path_str.split(' > ')
+                for family in reversed(family_path):
+                    if family in self.DESIRED_SUBFAMILIES:
+                        return family
+                return family_path[0]
+
+        logging.debug(f"Language code '{lang_code}' could not be resolved to a family.")
+        return "Unclassified"
+
+    def get_family_stats(self):
+        total = sum(self.family_counts.values())
+        if total == 0:
+            return {}
+        
+        stats = {}
+        for family, count in self.family_counts.items():
+            stats[family] = (count / total) * 100
+        return stats
 
 @cli.command()
 @click.argument('filenames', nargs=-1, required=True)
 @click.option('--allstats', is_flag=True,
         help="Get all etymological statistics about the file(s).")
 @click.option('--lang', default='eng',
-        help="Specify the language of the texts. Use ISO639-3 "\
+        help="Specify the language of the texts. Use ISO639-3 "
              "three-letter language code. Default is English.")
-@click.option('--showfamilies', help="A comma-separated list of language "\
+@click.option('--showfamilies', help="A comma-separated list of language "
               "families to show, e.g. Latinate,Germanic")
-@click.option('--affixes', is_flag=True, help="Don't ignore affixes. "\
+@click.option('--affixes', is_flag=True, help="Don't ignore affixes. "
               "Default is to ignore them.")
-@click.option('--current', is_flag=True, help="Don't ignore current language "\
+@click.option('--current', is_flag=True, help="Don't ignore current language "
               "and its middle variants. Default is to ignore them.")
 @click.option('-c', '--csv', is_flag=True, help="Print a machine-readable "
               "CSV instead of a pretty table.")
-@click.option('--chart', is_flag=True, help="Make a pretty graph of the "\
+@click.option('--chart', is_flag=True, help="Make a pretty graph of the "
               "results. For one text, a pie; for multiple, a bar.")
-@click.option('--verbose', is_flag=True, help="Show debugging messages.")
+@click.option('--debug', default='warning',
+              type=click.Choice(['debug', 'info', 'warning', 'error', 'critical']),
+              help="Set the logging level.")
+@click.option('--verbose', is_flag=True, help="Enable verbose output (alias for --debug=debug).", hidden=True)
 def analyze(filenames, allstats, lang, showfamilies, affixes,
-        current, csv, chart, verbose):
+        current, csv, chart, debug, verbose):
     """
     Analyzes a text(s) for the etymologies of its words, and tallies the words
     by origin language, and origin language family.
     """
-    single = len(filenames) == 1
-    ignoreAffixes = not affixes
-    ignoreCurrent = not current
-    cumulativeStats = {}
-    cumulativeAllStats = {}
-
     if verbose:
-        logging.basicConfig(level=logging.DEBUG)
+        loglevel = "debug"
+    else:
+        loglevel = debug
+    
+    logging.basicConfig(level=getattr(logging, loglevel.upper()))
+
+    db_path = Path.home() / ".local" / "share" / "macroetym" / "etymology.sqlite"
+    if not db_path.exists():
+        print(f"Database not found at {db_path}.")
+        print("Please run `macroetym init` to create the database.")
+        return
+
+    # Load the on-disk database into an in-memory database for speed.
+    disk_conn = sqlite3.connect(db_path)
+    mem_conn = sqlite3.connect(':memory:')
+    disk_conn.backup(mem_conn)
+    disk_conn.close()
+
+    cumulative_stats = {}
 
     for filename in filenames:
         try:
             with open(filename) as fdata:
                 text = fdata.read()
-        except UnicodeDecodeError as e:
-            logging.error(f"Can't open file {filename} in the normal way, using utf-8.")
-            logging.error(f"I'll try opening using other encodings, but you may want to Try converting it to utf-8..")
-            logging.error(f"Here's the error: {e}")
+        except UnicodeDecodeError:
+            logging.warning(f"Can't open file {filename} as UTF-8. Trying other encodings...")
             try:
                 with open(filename, encoding='utf-16') as fdata:
                     text = fdata.read()
-            except UnicodeDecodeError as e:
-                logging.error(f"Can't open file {filename} as UTF-16, either. Try converting it to utf-8. Error: {e}")
+            except UnicodeDecodeError:
+                logging.warning(f"Can't open file {filename} as UTF-16. Trying latin1...")
                 try:
                     with open(filename, encoding='latin1') as fdata:
                         text = fdata.read()
-                except UnicodeDecodeError as e:
-                    logging.error(f"Can't open file {filename} as latin1, either. Try converting it to utf-8. Error: {e}")
+                except UnicodeDecodeError:
+                    logging.error(f"Can't open file {filename} with any of the attempted encodings. Skipping.")
+                    continue
+        
+        analysis = Analysis(text, lang, mem_conn)
+        cumulative_stats[filename] = analysis.get_family_stats()
 
-        t = Text(text, lang, ignoreAffixes, ignoreCurrent)
+    mem_conn.close()
 
-        if single and allstats and csv:
-            t.printCSVStats(filename)
-        elif single and allstats:
-            t.printPrettyStats(filename)
-
-        if chart and single:
-            s = pd.Series(t.familyStats())
-            if showfamilies:
-                famlist = showfamilies.split(',')
-                s = s.loc[famlist]
-            ax = s.plot(kind='pie', figsize=(6,6))
-            ax.set_ylabel('') # Don't write the series name "None" on the left.
-            fig = ax.get_figure()
-            fig.savefig('chart.png')
-            print('Chart saved as chart.png.')
-
-        cumulativeStats[filename] = t.familyStats(pad=single)
-        cumulativeAllStats[filename] = t.prettyStats
-
-    df = pd.DataFrame(cumulativeStats)
+    df = pd.DataFrame(cumulative_stats)
     df = df.fillna(0)
-
-    dfAll = pd.DataFrame(cumulativeAllStats)
-    dfAll = dfAll.fillna(0)
 
     if showfamilies:
         famlist = showfamilies.split(',')
         df = df.loc[famlist]
 
-    if not allstats:
-        if csv:
-            print(df.to_csv())
-        else:
-            print(df)
+    if csv:
+        print(df.to_csv())
     else:
-        if csv:
-            print(dfAll.to_csv())
-        else:
-            print(dfAll)
+        print(df)
 
-    if chart and not single:
-        ax = df.plot(kind='bar', figsize=(6,6))
+    if chart:
+        if len(filenames) == 1:
+            ax = df.plot(kind='pie', y=filenames[0], figsize=(6,6))
+            ax.set_ylabel('')
+        else:
+            ax = df.plot(kind='bar', figsize=(6,6))
+        
         fig = ax.get_figure()
         fig.tight_layout()
         fig.savefig('chart.png')
@@ -478,16 +387,7 @@ def init(limit):
     Initialize the etymological database from Kaikki data.
     This will download a large file (several GB).
     """
-    with tempfile.TemporaryDirectory() as temp_dir_str:
-        temp_dir = Path(temp_dir_str)
-        print(f"Using temporary directory: {temp_dir}")
-        
-        downloaded_file_path = datamaker.download_kaikki_data(temp_dir)
-        
-        if downloaded_file_path:
-            datamaker.create_database(downloaded_file_path, limit=limit)
-        else:
-            print("Database creation failed because the download did not complete.")
+    data_download.init_db(limit=limit)
 
 @cli.command()
 def web():